Initial import of pipecheck code
I'm extracting this from https://github.com/featherbread/xt since
I'd like to reuse it in other contexts.

For the most part, I've just copied the code and changed the visibility
modifiers. I'd like to re-edit the doc comment too before publishing.

I'd also like for v1.0.0 to be the initial publish. From then on, my
goal is to publish major releases if Rust standardizes new provided
methods on the Write trait that I want to include proper wrappers for.
It's possible that I'd change the underlying implementation, e.g.
switching from libc to nix's signal functions, but it's not like there
will ever be new user-facing functionality in this crate.