Remove Cargo.lock
I know this is no longer the recommendation for library crates, but my
lockfile uses a version that Rust 1.36.0 can't read, so I can't even
`cargo check` it on that toolchain.

On top of that, I have to download the x86_64 toolchain to my M1 Mac,
since apparently 1.36 predates that architecture.

Even more on top of that, 1.36 super-predates Cargo not having to pull
an index of every crate in existence from a GitHub repo.

Hmm. Should I give up on having the minimum possible MSRV?